+++ src/tree.rs
@@ -1,69 +1,136 @@
-use miette::{diagnostic, IntoDiagnostic, Result, Severity};
 use std::{
     fmt::Display,
     path::{Path, PathBuf},
     sync::{Arc, Mutex, OnceLock, Weak},
 };
 
-use crate::gitignore::GitIgnore;
+use crate::{gitignore::GitIgnore, logger::Logger};
 
 const IS_CHILD_GLIPH: &str = "├─ ";
-const FIRST_CHILD_GLIPH: &str = "┬─ ";
 const LAST_CHILD_GLIPH: &str = "└─ ";
 const SKIP_GLIPH: &str = "│  ";
 const SKIP_GLIPH_GAP: &str = "   ";
 
+/// Represents a tree structure for storing hierarchical data.
 #[derive(Debug, Clone)]
 pub struct Tree<T: Clone + PartialEq + Display> {
+    /// A weak reference to this tree node.
+    _weak_self: Weak<Self>,
+    /// The current directory path of this tree node.
     current_dir: PathBuf,
+    /// A weak reference to the parent tree node.
     parent: Option<Weak<Tree<T>>>,
+    /// The leaves (files) of this tree node.
     leaves: Arc<Mutex<Vec<T>>>,
-    gitignore: OnceLock<GitIgnore>,
-    branches: Arc<Mutex<Vec<Tree<T>>>>,
+    /// The GitIgnore instance for this tree node, if any.
+    /// If `Some`, it means that the GitIgnore is a leaf in
+    /// the current branch (tree node).
+    gitignore: Arc<OnceLock<GitIgnore>>,
+    /// The child branches (directories) of this tree node.
+    branches: Arc<Mutex<Vec<Arc<Tree<T>>>>>,
 }
 
 impl<T: Clone + PartialEq + Display> Tree<T> {
-    pub fn new(current_dir: PathBuf, parent: Option<Weak<Tree<T>>>) -> Self {
-        Self {
+    /// Creates a new Tree instance.
+    ///
+    /// # Arguments
+    ///
+    /// * `current_dir` - The current directory path for this tree node.
+    /// * `parent` - An optional weak reference to the parent tree node.
+    ///
+    /// # Returns
+    ///
+    /// A new Tree instance.
+    pub fn new(current_dir: PathBuf, parent: Option<Weak<Tree<T>>>) -> Arc<Self> {
+        Arc::new_cyclic(|weak_self| Self {
+            _weak_self: weak_self.clone(),
             current_dir,
             parent,
             leaves: Arc::new(Mutex::new(Vec::new())),
-            gitignore: OnceLock::new(),
+            gitignore: Arc::new(OnceLock::new()),
             branches: Arc::new(Mutex::new(Vec::new())),
-        }
+        })
     }
 
+    /// Adds a leaf (file) to the tree.
+    ///
+    /// # Arguments
+    ///
+    /// * `leaf` - The leaf to add.
     pub fn add_leaf(&self, leaf: T) {
-        let mut leaves = self.leaves.lock().unwrap();
+        let mut leaves = self.leaves.lock().expect("Failed to lock leaves mutex");
         leaves.push(leaf);
     }
 
-    pub fn add_branch(&self, branch: Tree<T>) {
-        let mut branches = self.branches.lock().unwrap();
+    /// Adds a branch (directory) to the tree.
+    ///
+    /// # Arguments
+    ///
+    /// * `branch` - The branch to add.
+    pub fn add_branch(&self, branch: Arc<Tree<T>>) {
+        let mut branches = self.branches.lock().expect("Failed to lock branches mutex");
         branches.push(branch);
     }
 
+    /// Returns the current directory path of this tree node.
     pub fn current_dir(&self) -> &Path {
         &self.current_dir
     }
 
+    /// Returns the parent tree node, if any.
     pub fn parent(&self) -> Option<Arc<Tree<T>>> {
-        self.parent.as_ref().map(|parent| parent.upgrade().unwrap())
+        self.parent.as_ref().and_then(|parent| parent.upgrade())
     }
 
+    /// Returns the first GitIgnore instance that applies to this tree node.
     pub fn gitignore(&self) -> Option<GitIgnore> {
         if self.gitignore.get().is_some() {
+            Logger::trace(
+                format!(
+                    "GitIgnore found for current tree node: {}",
+                    self.current_dir.display()
+                )
+                .as_str(),
+            );
             return self.gitignore.get().cloned();
         }
         if let Some(parent) = self.parent() {
+            Logger::trace(
+                format!(
+                    "No GitIgnore found for current tree node: {}, checking parent",
+                    self.current_dir.display()
+                )
+                .as_str(),
+            );
             return parent.gitignore();
         }
+        Logger::trace(format!("Exhausted all parent nodes to find GitIgnore, ending search at current tree node: {}", self.current_dir.display()).as_str());
         None
     }
 
-    pub fn backtrack_to_branch(&self, path: &Path) -> Option<Tree<T>> {
+    /// Sets the GitIgnore instance for this tree node.
+    ///
+    /// # Arguments
+    ///
+    /// * `gitignore` - The GitIgnore instance to set.
+    pub fn set_gitignore(&self, gitignore: GitIgnore) {
+        self.gitignore
+            .set(gitignore)
+            .expect("Failed to set GitIgnore");
+    }
+
+    /// Backtracks to find the branch (directory) that contains the given path.
+    ///
+    /// # Arguments
+    ///
+    /// * `path` - The path to search for.
+    ///
+    /// # Returns
+    ///
+    /// The Tree instance that contains the given path, or None if not found.
+    pub fn backtrack_to_branch(&self, path: &Path) -> Option<Arc<Tree<T>>> {
         if path == self.current_dir() {
-            return Some(self.clone());
+            return Some(self._weak_self.upgrade().unwrap());
         }
         if let Some(parent) = self.parent() {
             return parent.backtrack_to_branch(path);
@@ -71,6 +138,11 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
         None
     }
 
+    /// Collects all leaves (files) from this tree node and its branches.
+    ///
+    /// # Returns
+    ///
+    /// A vector containing all leaves in the tree.
     pub fn collect_all_leaves(&self) -> Vec<T> {
         let mut local_leaves = {
             let local_leaves_lock = self.leaves.lock().unwrap();
@@ -87,6 +159,29 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
         local_leaves
     }
 
+    /// Collects all leaves (files) at this tree node.
+    ///
+    /// # Returns
+    ///
+    /// A vector containing all leaves in the tree node.
+    pub fn collect_local_leaves(&self) -> Vec<T> {
+        self.leaves.lock().unwrap().clone()
+    }
+
+    /// Collects all branches (directories) at this tree node.
+    ///
+    /// # Returns
+    ///
+    /// A vector containing all branches in the tree node.
+    pub fn collect_local_branches(&self) -> Vec<Arc<Tree<T>>> {
+        self.branches.lock().unwrap().clone()
+    }
+
+    /// Generates a string representation of the tree structure.
+    ///
+    /// # Returns
+    ///
+    /// A string representing the tree structure.
     pub fn to_string(&self) -> String {
         let mut buffer = String::new();
         self.build_string(&mut buffer, "", true);
@@ -95,6 +190,13 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
         buffer
     }
 
+    /// Helper method to recursively build the string representation of the tree.
+    ///
+    /// # Arguments
+    ///
+    /// * `buffer` - The string buffer to append to.
+    /// * `prefix` - The prefix to use for the current line.
+    /// * `is_last` - Whether this is the last item in the current level.
     fn build_string(&self, buffer: &mut String, branch_prefix: &str, is_last_at_level: bool) {
         let branches_len = self.branches.lock().unwrap().len();
         let leaves_len = self.leaves.lock().unwrap().len();
@@ -121,7 +223,7 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
                 branch_prefix,
                 if dir_name.is_none() {
                     ""
-                } else if i == branches_len - 1 && leaves_len == 0 {
+                } else if i == branches_len - 1 && leaves_len == 0 && is_last_at_level {
                     SKIP_GLIPH_GAP
                 } else {
                     SKIP_GLIPH
@@ -136,17 +238,9 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
         }
 
         for (i, leaf) in self.leaves.lock().unwrap().iter().enumerate() {
-            let new_leaf_prefix = format!(
-                "{}{}",
-                branch_prefix,
-                if i == leaves_len - 1 {
-                    LAST_CHILD_GLIPH
-                } else {
-                    IS_CHILD_GLIPH
-                }
-            );
             let new_leaf_display = format!(
-                "{}{}{}\n",
+                "{}{}{}{}\n",
+                branch_prefix,
                 if dir_name.is_none() {
                     ""
                 } else if !is_last_at_level {
@@ -154,7 +248,11 @@ impl<T: Clone + PartialEq + Display> Tree<T> {
                 } else {
                     SKIP_GLIPH_GAP
                 },
-                new_leaf_prefix,
+                if i == leaves_len - 1 {
+                    LAST_CHILD_GLIPH
+                } else {
+                    IS_CHILD_GLIPH
+                },
                 leaf.to_string()
             );
 
@@ -182,7 +280,7 @@ mod tests {
     #[test]
     fn test_new_tree() {
         let root_path = PathBuf::from("/");
-        let tree: Tree<String> = Tree::new(root_path.clone(), None);
+        let tree: Arc<Tree<String>> = Tree::new(root_path.clone(), None);
 
         assert_eq!(tree.current_dir(), root_path.as_path());
         assert!(tree.parent().is_none());
@@ -193,7 +291,7 @@ mod tests {
     #[test]
     fn test_add_leaf() {
         let root_path = PathBuf::from("/");
-        let tree: Tree<String> = Tree::new(root_path, None);
+        let tree: Arc<Tree<String>> = Tree::new(root_path, None);
 
         tree.add_leaf(String::from("leaf1"));
         assert_eq!(tree.leaves.lock().unwrap().len(), 1);
@@ -203,13 +301,10 @@ mod tests {
     #[test]
     fn test_add_branch() {
         let root_path = PathBuf::from("/");
-        let tree: Tree<String> = Tree::new(root_path.clone(), None);
+        let tree: Arc<Tree<String>> = Tree::new(root_path.clone(), None);
 
         let branch_path = PathBuf::from("/branch");
-        let branch: Tree<String> = Tree::new(
-            branch_path.clone(),
-            Some(Arc::downgrade(&Arc::new(tree.clone()))),
-        );
+        let branch: Arc<Tree<String>> = Tree::new(branch_path.clone(), Some(Arc::downgrade(&tree)));
 
         tree.add_branch(branch.clone());
         assert_eq!(tree.branches.lock().unwrap().len(), 1);
@@ -222,12 +317,12 @@ mod tests {
     #[test]
     fn test_parent() {
         let root_path = PathBuf::from("/");
-        let tree: Tree<String> = Tree::new(root_path.clone(), None);
+        let tree: Arc<Tree<String>> = Tree::new(root_path.clone(), None);
 
         let root = Arc::new(tree);
 
         let branch_path = PathBuf::from("/branch");
-        let branch: Tree<String> = Tree::new(branch_path.clone(), Some(Arc::downgrade(&root)));
+        let branch: Arc<Tree<String>> = Tree::new(branch_path.clone(), Some(Arc::downgrade(&root)));
 
         assert!(branch.parent().is_some());
         assert_eq!(branch.parent().unwrap().current_dir(), root_path.as_path());
@@ -236,16 +331,16 @@ mod tests {
     #[test]
     fn test_backtrack_to_branch() {
         let root_path = PathBuf::from("/");
-        let root: Arc<Tree<String>> = Arc::new(Tree::new(root_path.clone(), None));
+        let root: Arc<Arc<Tree<String>>> = Arc::new(Tree::new(root_path.clone(), None));
 
         let branch_path = PathBuf::from("/branch");
-        let branch: Arc<Tree<String>> = Arc::new(Tree::new(
+        let branch: Arc<Arc<Tree<String>>> = Arc::new(Tree::new(
             branch_path.clone(),
             Some(Arc::downgrade(&root.clone())),
         ));
 
         let leaf_path = PathBuf::from("/branch/leaf");
-        let leaf: Tree<String> =
+        let leaf: Arc<Tree<String>> =
             Tree::new(leaf_path.clone(), Some(Arc::downgrade(&branch.clone())));
 
         assert_eq!(
@@ -305,7 +400,8 @@ mod tests {
             .expect("Failed to create GitIgnore")
             .expect("GitIgnore is None");
 
-        let tree1: Arc<Tree<String>> = Arc::new(Tree::new(temp_dir.path().to_path_buf(), None));
+        let tree1: Arc<Arc<Tree<String>>> =
+            Arc::new(Tree::new(temp_dir.path().to_path_buf(), None));
         tree1.gitignore.set(gitignore.clone()).unwrap();
 
         let branch_path = temp_dir.path().join("branch");
@@ -326,7 +422,8 @@ mod tests {
             .expect("Failed to create GitIgnore")
             .expect("GitIgnore is None");
 
-        let tree1: Arc<Tree<String>> = Arc::new(Tree::new(temp_dir.path().to_path_buf(), None));
+        let tree1: Arc<Arc<Tree<String>>> =
+            Arc::new(Tree::new(temp_dir.path().to_path_buf(), None));
         tree1.gitignore.set(gitignore.clone()).unwrap();
 
         let branch_path = temp_dir.path().join("branch");